style(crates,tests): apply rustfmt standardizations and clean clippy lints across workspace
- Standardize formatting and formatting comments across fsci-conformance tests,
  fsci-cluster, fsci-linalg, fsci-sparse, fsci-interpolate, and fsci-ndimage.
- Clean up unused imports, dead assignments, and float formatting in tests.

+++ crates/fsci-conformance/tests/diff_interpolate_pade.rs
@@ -119,7 +119,7 @@ fn generate_query() -> OracleQuery {
             if k == 0 {
                 0.0
             } else {
-                (-1.0_f64).powi((k + 1) as i32) / (k as f64)
+                (-1.0_f64).powi(k + 1) / (k as f64)
             }
         })
         .collect();
